refactor: update entity imports and relationships for consistency
- Changed entity imports in session.entity.ts and todo.entity.ts to use direct class references, improving clarity and preventing minification issues.
- Updated ManyToOne and OneToMany relationships in User, Session, and Todo entities to utilize class references for better compatibility with MikroORM.
